Add and configure some development tools
This adds Guard, RuboCop, and Foreman to the project along with their
associated configurations.

Guard watches for file changes and automatically runs tasks. Thus far
tasks for running tests and RuboCop are configured.

RuboCop is a Ruby linter and style checker. Configuration for it is
found in the various `.rubocop.yml` files in different directories.
(RuboCop walks the directory hierarchy for each file it checks looking
for a config and uses the first one it finds.) By default, some checks
that clash with Rails and autogenerated files (such as `db/schema.rb`)
are disabled.

Foreman will automatically run multiple processes as specified in the
`procfile`. Currently it is configured to run the rails server and
guard.